+++ wiki/CHANGES.md
@@ -2,6 +2,9 @@
 
 Check also [MIGRATION.md](MIGRATION.md) for possible compatibility problems.
 
+### 0.8.1
+* [#245] Fixed: back to 'jar' packaging
+
 ### 0.8.0
 * [#244] Added: `StreamEx.toMutableList` and `StreamEx.toMutableSet`.  
  Removed mutability guarantee for `toList` and `toSet`.